// src/types/settings.ts
export interface SiteSettings {
  id: string
  companyName?: string
  companyEmail?: string
  companyPhone?: string
  companyAddress?: string
  companyCity?: string
  companyState?: string
  companyZipCode?: string
  websiteUrl?: string
  logoUrl?: string
  colorPrimary?: string
  colorSecondary?: string
  termsOfService?: string
  privacyPolicy?: string
  createdAt: string
  updatedAt: string
}